requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
faker>=22.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0
tqdm>=4.66.0